+++ crates/forge/src/main.rs
@@ -10,12 +10,11 @@ use forge::run;
 use forge::{pretty_printing, RunnerConfig};
 
 use forge::scarb::{
-    dependencies_for_package, get_contracts_map, paths_for_package, target_name_for_package,
-    try_get_starknet_artifacts_path,
+    corelib_for_package, dependencies_for_package, get_contracts_map, paths_for_package,
+    target_name_for_package, try_get_starknet_artifacts_path,
 };
 use std::process::Command;
 
-static CORELIB_PATH: Dir = include_dir!("./corelib/src");
 static PREDEPLOYED_CONTRACTS: Dir = include_dir!("crates/cheatable-starknet/predeployed-contracts");
 
 #[derive(Parser, Debug)]
@@ -32,14 +31,6 @@ struct Args {
     exit_first: bool,
 }
 
-fn load_corelib() -> Result<TempDir> {
-    let tmp_dir = tempdir()?;
-    CORELIB_PATH
-        .extract(&tmp_dir)
-        .context("Failed to copy corelib to temporary directory")?;
-    Ok(tmp_dir)
-}
-
 fn load_predeployed_contracts() -> Result<TempDir> {
     let tmp_dir = tempdir()?;
     PREDEPLOYED_CONTRACTS
@@ -50,12 +41,6 @@ fn load_predeployed_contracts() -> Result<TempDir> {
 
 fn main_execution() -> Result<()> {
     let args = Args::parse();
-
-    // TODO #1997
-    let corelib_dir = load_corelib()?;
-    let corelib_path: PathBuf = corelib_dir.path().into();
-    let corelib = Utf8PathBuf::try_from(corelib_path)
-        .context("Failed to convert corelib path to Utf8PathBuf")?;
 
     let predeployed_contracts_dir = load_predeployed_contracts()?;
     let predeployed_contracts_path: PathBuf = predeployed_contracts_dir.path().into();
@@ -85,6 +70,7 @@ fn main_execution() -> Result<()> {
         let (package_path, lib_path) = paths_for_package(&scarb_metadata, package)?;
         let dependencies = dependencies_for_package(&scarb_metadata, package)?;
         let target_name = target_name_for_package(&scarb_metadata, package)?;
+        let corelib_path = corelib_for_package(&scarb_metadata, package)?;
         let runner_config = RunnerConfig::new(
             args.test_name.clone(),
             args.exact,
@@ -103,7 +89,7 @@ fn main_execution() -> Result<()> {
             &lib_path,
             &Some(dependencies.clone()),
             &runner_config,
-            Some(&corelib),
+            &corelib_path,
             &contracts,
             &predeployed_contracts,
         )?;