+++ crates/forge/tests/integration/deploy.rs
@@ -1,4 +1,4 @@
-use crate::integration::common::corelib::{corelib, predeployed_contracts};
+use crate::integration::common::corelib::{corelib_path, predeployed_contracts};
 use crate::integration::common::runner::Contract;
 use crate::{assert_case_output_contains, assert_failed, assert_passed, test_case};
 use camino::Utf8PathBuf;
@@ -12,8 +12,7 @@ fn error_handling() {
         indoc!(
             r#"
         use result::ResultTrait;
-        use cheatcodes::RevertedTransactionTrait;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy };
         use array::ArrayTrait;
         
         #[test]
@@ -63,8 +62,8 @@ fn error_handling() {
         &test.path().unwrap().join("src/lib.cairo"),
         &Some(test.linked_libraries()),
         &Default::default(),
-        Some(&Utf8PathBuf::from_path_buf(corelib().to_path_buf()).unwrap()),
-        &test.contracts(corelib().path()).unwrap(),
+        &corelib_path(),
+        &test.contracts(&corelib_path()).unwrap(),
         &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
     )
     .unwrap();
@@ -78,8 +77,7 @@ fn deploy_fails_on_calldata_when_contract_has_no_constructor() {
         indoc!(
             r#"
         use result::ResultTrait;
-        use cheatcodes::RevertedTransactionTrait;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy };
         use array::ArrayTrait; 
             
         #[test]
@@ -108,8 +106,8 @@ fn deploy_fails_on_calldata_when_contract_has_no_constructor() {
         &test.path().unwrap().join("src/lib.cairo"),
         &Some(test.linked_libraries()),
         &Default::default(),
-        Some(&Utf8PathBuf::from_path_buf(corelib().to_path_buf()).unwrap()),
-        &test.contracts(corelib().path()).unwrap(),
+        &corelib_path(),
+        &test.contracts(&corelib_path()).unwrap(),
         &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
     )
     .unwrap();
@@ -123,8 +121,7 @@ fn test_deploy_fails_on_missing_constructor_arguments() {
         indoc!(
             r#"
         use result::ResultTrait;
-        use cheatcodes::RevertedTransactionTrait;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy };
         use array::ArrayTrait; 
             
         #[test]
@@ -161,8 +158,8 @@ fn test_deploy_fails_on_missing_constructor_arguments() {
         &test.path().unwrap().join("src/lib.cairo"),
         &Some(test.linked_libraries()),
         &Default::default(),
-        Some(&Utf8PathBuf::from_path_buf(corelib().to_path_buf()).unwrap()),
-        &test.contracts(corelib().path()).unwrap(),
+        &corelib_path(),
+        &test.contracts(&corelib_path()).unwrap(),
         &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
     )
     .unwrap();
@@ -176,8 +173,7 @@ fn test_deploy_fails_on_too_many_constructor_arguments() {
         indoc!(
             r#"
         use result::ResultTrait;
-        use cheatcodes::RevertedTransactionTrait;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy };
         use array::ArrayTrait;
 
         #[test]
@@ -219,8 +215,8 @@ fn test_deploy_fails_on_too_many_constructor_arguments() {
         &test.path().unwrap().join("src/lib.cairo"),
         &Some(test.linked_libraries()),
         &Default::default(),
-        Some(&Utf8PathBuf::from_path_buf(corelib().to_path_buf()).unwrap()),
-        &test.contracts(corelib().path()).unwrap(),
+        &corelib_path(),
+        &test.contracts(&corelib_path()).unwrap(),
         &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
     )
     .unwrap();
@@ -235,8 +231,7 @@ fn test_deploy_fails_with_incorrect_class_hash() {
             r#"
         use result::ResultTrait;
         use option::OptionTrait;
-        use cheatcodes::RevertedTransactionTrait;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy };
         use array::ArrayTrait;
         use traits::TryInto;
         use starknet::Felt252TryIntoClassHash;
@@ -272,8 +267,8 @@ fn test_deploy_fails_with_incorrect_class_hash() {
         &test.path().unwrap().join("src/lib.cairo"),
         &Some(test.linked_libraries()),
         &Default::default(),
-        Some(&Utf8PathBuf::from_path_buf(corelib().to_path_buf()).unwrap()),
-        &test.contracts(corelib().path()).unwrap(),
+        &corelib_path(),
+        &test.contracts(&corelib_path()).unwrap(),
         &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
     )
     .unwrap();
@@ -289,8 +284,7 @@ fn test_deploy_invokes_the_constructor() {
             r#"
         use option::OptionTrait;
         use result::ResultTrait;
-        use cheatcodes::RevertedTransactionTrait;
-        use cheatcodes::PreparedContract;
+        use cheatcodes::{ declare, PreparedContract, deploy };
         use array::ArrayTrait;
         use traits::TryInto;
         use starknet::ContractAddress;
@@ -355,8 +349,8 @@ fn test_deploy_invokes_the_constructor() {
         &test.path().unwrap().join("src/lib.cairo"),
         &Some(test.linked_libraries()),
         &Default::default(),
-        Some(&Utf8PathBuf::from_path_buf(corelib().to_path_buf()).unwrap()),
-        &test.contracts(corelib().path()).unwrap(),
+        &corelib_path(),
+        &test.contracts(&corelib_path()).unwrap(),
         &Utf8PathBuf::from_path_buf(predeployed_contracts().to_path_buf()).unwrap(),
     )
     .unwrap();