+++ CHANGELOG.md
@@ -15,6 +15,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Changed
 
+- Exported cheatcodes as a Scarb package, now you have to import them explicitly with e.g. `use cheatcodes::declare`
+  and add them as a dependency to your Scarb.toml
+
+```toml
+[dependencies]
+# ...
+cheatcodes = { git = "https://github.com/foundry-rs/starknet-foundry", tag = "0.4.0" }
+```
+
 - Moved `ForgeConfigFromScarb` to `scarb.rs` and renamed to `ForgeConfig`
 - Made private:
     - `print_collected_tests_count`